+++ fail/fail_test.go
@@ -3,5 +3,7 @@ package fail
 import "testing"
 
 func TestFail(t *testing.T) {
+	// TODO: MismatchCount
 	// TODO: MissingParams
+	// TODO: NotFoundItem
 }