+++ model/schema_combo_test.go
@@ -1,0 +1,13 @@
+package model
+
+import "testing"
+
+func TestCombo(t *testing.T) {
+	// TODO: Schema.GetOrCreate
+	// TODO: Schema.GetOrCreateTx
+	// TODO: Schema.GetAndLockTx
+	// TODO: Schema.GetAndLockTxItems
+	// TODO: Schema.GetOrCreateAndLockTx
+	// TODO: Schema.UpdateAndGetTx
+	// TODO: MoveItemTx
+}